0004-MEDIAN-OF-TWO-SORTED-ARRAYS Stats: Time: 0 ms (100%), Space: 12.6 MB (51.71%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.py
@@ -8,5 +8,4 @@ class Solution(object):
         list1=nums1+nums2
         list1.sort()
         n=len(list1)
-        return list1[n/2] if n%2!=0 else float(float((list1[n/2])+float(list1[(n/2)-1]))/2)
-        
+        return list1[n/2] if n%2!=0 else (float(list1[n/2])+float(list1[n/2-1]))/2
